requirements.txt
ipykernel
CurrencyConverter
duckduckgo-search
python-dotenv
langchain
langchain-groq
langchain-community
langgraph
streamlit
-e .